fix(nextjs): Make all wrappers isomorphic and available in all runtimes (#8743)
