--- packages/nextjs/src/client/wrapAppGetInitialPropsWithSentry.ts
@@ -1,20 +1,0 @@
-import type App from 'next/app';
-
-type AppGetInitialProps = (typeof App)['getInitialProps'];
-
-/**
- * A passthrough function in case this function is used on the clientside. We need to make the returned function async
- * so we are consistent with the serverside implementation.
- */
-export function wrapAppGetInitialPropsWithSentry(origAppGetInitialProps: AppGetInitialProps): AppGetInitialProps {
-  return new Proxy(origAppGetInitialProps, {
-    apply: (wrappingTarget, thisArg, args: Parameters<AppGetInitialProps>) => {
-      return wrappingTarget.apply(thisArg, args);
-    },
-  });
-}
-
-/**
- * @deprecated Use `wrapAppGetInitialPropsWithSentry` instead.
- */
-export const withSentryServerSideAppGetInitialProps = wrapAppGetInitialPropsWithSentry;
