+++ packages/nextjs/src/common/utils/platformSupportsStreaming.ts
