+++ packages/nextjs/src/config/templates/apiWrapperTemplate.ts
@@ -13,10 +13,7 @@ import * as origModule from '__SENTRY_WRAPPING_TARGET_FILE__';
 import * as Sentry from '@sentry/nextjs';
 import type { PageConfig } from 'next';
 
-import type { VercelCronsConfig } from '../../common/types';
-// We import this from `wrappers` rather than directly from `next` because our version can work simultaneously with
-// multiple versions of next. See note in `wrappers/types` for more.
-import type { NextApiHandler } from '../../server/types';
+import type { NextApiHandler, VercelCronsConfig } from '../../common/types';
 
 type NextApiModule = (
   | {
