+++ packages/nextjs/src/edge/wrapApiHandlerWithSentry.ts
@@ -1,7 +1,7 @@
 import { getCurrentHub } from '@sentry/core';
 
+import { withEdgeWrapping } from '../common/utils/edgeWrapperUtils';
 import type { EdgeRouteHandler } from './types';
-import { withEdgeWrapping } from './utils/edgeWrapperUtils';
 
 /**
  * Wraps a Next.js edge route handler with Sentry error and performance instrumentation.
